+++ src/test-integration/resources/test-instrumented-file-system.js
@@ -1,14 +1,12 @@
 function BranchData() {
     this.position = -1;
     this.nodeLength = -1;
-    this.src = null;
     this.evalFalse = 0;
     this.evalTrue = 0;
 
-    this.init = function(position, nodeLength, src) {
+    this.init = function(position, nodeLength) {
         this.position = position;
         this.nodeLength = nodeLength;
-        this.src = src;
         return this;
     };
 
@@ -35,18 +33,17 @@ function BranchData() {
     this.toJSON = function() {
         return '{"position":' + this.position
             + ',"nodeLength":' + this.nodeLength
-            + ',"src":' + jscoverage_quote(this.src)
             + ',"evalFalse":' + this.evalFalse
             + ',"evalTrue":' + this.evalTrue + '}';
     };
 
-    this.message = function() {
+    this.message = function(src) {
         if (this.evalTrue === 0 && this.evalFalse === 0)
-            return 'Condition never evaluated         :\t' + this.src;
+            return 'Condition never evaluated         :\t' + src;
         else if (this.evalTrue === 0)
-            return 'Condition never evaluated to true :\t' + this.src;
+            return 'Condition never evaluated to true :\t' + src;
         else if (this.evalFalse === 0)
-            return 'Condition never evaluated to false:\t' + this.src;
+            return 'Condition never evaluated to false:\t' + src;
         else
             return 'Condition covered';
     };
@@ -55,7 +52,7 @@ function BranchData() {
 BranchData.fromJson = function(jsonString) {
     var json = eval('(' + jsonString + ')');
     var branchData = new BranchData();
-    branchData.init(json.position, json.nodeLength, json.src);
+    branchData.init(json.position, json.nodeLength);
     branchData.evalFalse = json.evalFalse;
     branchData.evalTrue = json.evalTrue;
     return branchData;
@@ -63,7 +60,7 @@ BranchData.fromJson = function(jsonString) {
 
 BranchData.fromJsonObject = function(json) {
     var branchData = new BranchData();
-    branchData.init(json.position, json.nodeLength, json.src);
+    branchData.init(json.position, json.nodeLength);
     branchData.evalFalse = json.evalFalse;
     branchData.evalTrue = json.evalTrue;
     return branchData;
@@ -74,7 +71,7 @@ function buildBranchMessage(conditions) {
     var i;
     for (i = 0; i < conditions.length; i++) {
         if (conditions[i] !== undefined && conditions[i] !== null && !conditions[i].covered())
-            message += '\n- '+ conditions[i].message();
+            message += '\n- '+ conditions[i].message(conditions[i].src);
     }
     return message;
 }