+++ src/test/java/jscover/report/lcov/LCovGeneratorTest.java
@@ -404,6 +404,6 @@ public class LCovGeneratorTest {
     }
 
     private BranchData getBranchData(int evalFalse, int evalTrue) {
-        return new BranchData(0, 0, null, evalFalse, evalTrue);
+        return new BranchData(0, 0, evalFalse, evalTrue);
     }
 }