+++ src/test/javascript/spec/branch_data_spec.js
@@ -27,49 +27,41 @@ describe('JSCover', function() {
 
         beforeEach(function() {
             branchData = new BranchData();
-            branchData.init(10,5,'x<y');
+            branchData.init(10,5);
         });
 
         it("should record position and length", function() {
             expect(branchData.position).toEqual(10);
             expect(branchData.nodeLength).toEqual(5);
-            expect(branchData.src).toEqual('x<y');
         });
 
         it("should not be covered if neither path evaluated", function() {
             expect(branchData.covered()).toBeFalsy();
             expect(branchData.pathsCovered()).toEqual(0);
-            expect(branchData.message()).toEqual('Condition never evaluated         :\tx<y');
+            expect(branchData.message('x<y')).toEqual('Condition never evaluated         :\tx<y');
         });
 
         it("should convert to and from JSON", function() {
             var json = BranchData.fromJson(branchData.toJSON());
             expect(json.position).toEqual(10);
             expect(json.nodeLength).toEqual(5);
-            expect(json.src).toEqual('x<y');
             expect(json.evalFalse).toEqual(0);
             expect(json.evalTrue).toEqual(0);
             expect(json.covered()).toEqual(false);
-        });
-
-        it("should escape source", function() {
-            branchData.src = '\'a\'.length + "b".length';
-            var json = BranchData.fromJson(branchData.toJSON());
-            expect(json.src).toEqual('\'a\'.length + "b".length');
         });
 
         it("should not be covered if only false path evaluated", function() {
             branchData.ranCondition(false);
             expect(branchData.covered()).toBeFalsy();
             expect(branchData.pathsCovered()).toEqual(1);
-            expect(branchData.message()).toEqual('Condition never evaluated to true :\tx<y');
+            expect(branchData.message('x<y')).toEqual('Condition never evaluated to true :\tx<y');
         });
 
         it("should not be covered if only true path evaluated", function() {
             branchData.ranCondition(true);
             expect(branchData.covered()).toBeFalsy();
             expect(branchData.pathsCovered()).toEqual(1);
-            expect(branchData.message()).toEqual('Condition never evaluated to false:\tx<y');
+            expect(branchData.message('x<y')).toEqual('Condition never evaluated to false:\tx<y');
         });
 
         it("should be covered if both paths evaluated", function() {
@@ -77,7 +69,7 @@ describe('JSCover', function() {
             branchData.ranCondition(true);
             expect(branchData.covered()).toBeTruthy();
             expect(branchData.pathsCovered()).toEqual(2);
-            expect(branchData.message()).toEqual('Condition covered');
+            expect(branchData.message('x<y')).toEqual('Condition covered');
         });
 
         it("should be covered if both paths evaluated multiple times", function() {
@@ -89,7 +81,7 @@ describe('JSCover', function() {
             expect(branchData.evalFalse).toEqual(2);
             expect(branchData.evalTrue).toEqual(2);
             expect(branchData.pathsCovered()).toEqual(2);
-            expect(branchData.message()).toEqual('Condition covered');
+            expect(branchData.message('x<y')).toEqual('Condition covered');
         });
     });
 
@@ -97,20 +89,21 @@ describe('JSCover', function() {
 
         it("should convert multiple conditions to JSON", function() {
             var lineN = new Array();
-            lineN[1] = new BranchData().init(1,10,'src1');
-            lineN[2] = new BranchData().init(2,20,'src2');
+            lineN[1] = new BranchData().init(1,10);
+            lineN[2] = new BranchData().init(2,20);
             lineN[2].ranCondition(false);
             lineN[2].ranCondition(true);
 
             var json = convertBranchDataConditionArrayToJSON(lineN);
-            expect(json).toEqual('[null,{"position":1,"nodeLength":10,"src":"src1","evalFalse":0,"evalTrue":0},{"position":2,"nodeLength":20,"src":"src2","evalFalse":1,"evalTrue":1}]');
+            expect(json).toEqual('[null,{"position":1,"nodeLength":10,"evalFalse":0,"evalTrue":0},{"position":2,"nodeLength":20,"evalFalse":1,"evalTrue":1}]');
         });
 
         it("should build branch message for line", function() {
             var conditions = new Array();
-            conditions[1] = new BranchData().init(1,10,'src1');
-            conditions[2] = undefined;
-            conditions[2] = new BranchData().init(2,20,'src2');
+            conditions[1] = new BranchData().init(1,10);
+            conditions[1].src = 'src1';
+            conditions[2] = new BranchData().init(2,20);
+            conditions[2].src = 'src2';
 
             var message = buildBranchMessage(conditions);
             var expected = 'The following was not covered:\n- Condition never evaluated         :\tsrc1\n- Condition never evaluated         :\tsrc2';
@@ -128,11 +121,11 @@ describe('JSCover', function() {
         it("should convert multiple lines to JSON and back", function() {
             var lines = {};
             lines['1'] = new Array();
-            lines['1'][1] = new BranchData().init(1,10,'src1');
-            lines['1'][2] = new BranchData().init(2,20,'src2');
+            lines['1'][1] = new BranchData().init(1,10);
+            lines['1'][2] = new BranchData().init(2,20);
             lines['2'] = new Array();
-            lines['2'][1] = new BranchData().init(3,30,'src3');
-            lines['2'][2] = new BranchData().init(4,40,'src4');
+            lines['2'][1] = new BranchData().init(3,30);
+            lines['2'][2] = new BranchData().init(4,40);
             lines['2'][2].ranCondition(false);
             lines['2'][2].ranCondition(true);
 
@@ -161,11 +154,11 @@ describe('JSCover', function() {
         it("should ignore extra object properties", function() {
             var lines = {};
             lines['1'] = new Array();
-            lines['1'][1] = new BranchData().init(1,10,'src1');
+            lines['1'][1] = new BranchData().init(1,10);
             lines['weird'] = new Array();
 
             var json = convertBranchDataLinesToJSON(lines);
-            expect(json).toEqual('{"1":[null,{"position":1,"nodeLength":10,"src":"src1","evalFalse":0,"evalTrue":0}]}');
+            expect(json).toEqual('{"1":[null,{"position":1,"nodeLength":10,"evalFalse":0,"evalTrue":0}]}');
         });
     });
 });